docs(headless,headless-react): refactor typedoc to central location
