+++ packages/headless-react/typedoc.json
@@ -8,6 +8,6 @@
   "categorizeByGroup": true,
   "entryPoints": ["./temp/ssr-commerce.json", "./temp/ssr.json"],
   "gitRevision": "master",
-  "plugin": ["./typedoc/dist/index.js"],
+  "plugin": ["../shared-typedoc/dist/index.js"],
   "projectDocuments": ["./README.md"]
 }
