--- packages/headless-react/typedoc/tsconfig.json
@@ -1,9 +1,0 @@
-{
-  "extends": "../tsconfig.typedoc.json",
-  "compilerOptions": {
-    "outDir": "dist",
-    "rootDir": "lib",
-    "jsx": "react"
-  },
-  "include": ["lib/**/*"]
-}
