+++ packages/headless/package.json
@@ -136,16 +136,16 @@
     "publish:npm": "npm run-script -w=@coveo/ci npm-publish",
     "publish:bump": "npm run-script -w=@coveo/ci bump",
     "build:typedoc": "npm run build:typedoc:plugin && npm run build:typedoc:docs && npm run build:typedoc:merge",
-    "build:typedoc:plugin": "npx tsc -p typedoc/tsconfig.json",
+    "build:typedoc:plugin": "npx tsc -p ../shared-typedoc/tsconfig.json",
     "build:typedoc:docs": "concurrently \"npm run build:typedoc:search\" \"npm run build:typedoc:commerce\" \"npm run build:typedoc:case-assist\" \"npm run build:typedoc:insight\" \"npm run build:typedoc:recommendation\" \"npm run build:typedoc:ssr\" \"npm run build:typedoc:ssr-commerce\"",
     "build:typedoc:merge": "typedoc --tsconfig tsconfig.typedoc.json",
-    "build:typedoc:search": "typedoc --json ./temp/index.json --tsconfig tsconfig.typedoc.json --options ./typedoc/configs/index.typedoc.json",
-    "build:typedoc:commerce": "typedoc --json ./temp/commerce.json --tsconfig tsconfig.typedoc.json --options ./typedoc/configs/commerce.typedoc.json",
-    "build:typedoc:case-assist": "typedoc --json ./temp/case-assist.json --tsconfig tsconfig.typedoc.json --options ./typedoc/configs/case-assist.typedoc.json",
-    "build:typedoc:insight": "typedoc --json ./temp/insight.json --tsconfig tsconfig.typedoc.json --options ./typedoc/configs/insight.typedoc.json",
-    "build:typedoc:recommendation": "typedoc --json ./temp/recommendation.json --tsconfig tsconfig.typedoc.json --options ./typedoc/configs/recommendation.typedoc.json",
-    "build:typedoc:ssr": "typedoc --json ./temp/ssr.json --tsconfig tsconfig.typedoc.json --options ./typedoc/configs/ssr.typedoc.json",
-    "build:typedoc:ssr-commerce": "typedoc --json ./temp/ssr-commerce.json --tsconfig tsconfig.typedoc.json --options ./typedoc/configs/ssr-commerce.typedoc.json",
+    "build:typedoc:search": "typedoc --json ./temp/index.json --tsconfig tsconfig.typedoc.json --options ../shared-typedoc/configs/headless/index.typedoc.json",
+    "build:typedoc:commerce": "typedoc --json ./temp/commerce.json --tsconfig tsconfig.typedoc.json --options ../shared-typedoc/configs/headless/commerce.typedoc.json",
+    "build:typedoc:case-assist": "typedoc --json ./temp/case-assist.json --tsconfig tsconfig.typedoc.json --options ../shared-typedoc/configs/headless/case-assist.typedoc.json",
+    "build:typedoc:insight": "typedoc --json ./temp/insight.json --tsconfig tsconfig.typedoc.json --options ../shared-typedoc/configs/headless/insight.typedoc.json",
+    "build:typedoc:recommendation": "typedoc --json ./temp/recommendation.json --tsconfig tsconfig.typedoc.json --options ../shared-typedoc/configs/headless/recommendation.typedoc.json",
+    "build:typedoc:ssr": "typedoc --json ./temp/ssr.json --tsconfig tsconfig.typedoc.json --options ../shared-typedoc/configs/headless/ssr.typedoc.json",
+    "build:typedoc:ssr-commerce": "typedoc --json ./temp/ssr-commerce.json --tsconfig tsconfig.typedoc.json --options ../shared-typedoc/configs/headless/ssr-commerce.typedoc.json",
     "serve:typedoc": "live-server ./docs",
     "promote:npm:latest": "npm run-script -w=@coveo/ci promote-npm-prod"
   },
