--- packages/headless/typedoc/lib/insertMetaTags.ts
@@ -1,11 +1,0 @@
-export function insertMetaTags() {
-  const head = document.getElementsByTagName('head')[0];
-  if (head) {
-    head.innerHTML += `
-      <meta name="color-scheme" content="light dark">
-      <meta name="theme-color" content="light">
-      <meta name="docsSiteUrl" content="https://docs.coveo.com">
-      <meta name="docsSiteBaseUrl" content="/en">
-      `;
-  }
-}
