--- packages/headless/typedoc/package.json
@@ -1,6 +1,0 @@
-{
-  "type": "module",
-  "devDependencies": {
-    "typedoc": "0.28.3"
-  }
-}
