+++ packages/shared-typedoc/SEARCHREADME.md
