+++ packages/shared-typedoc/configs/headless-react/ssr-commerce.typedoc.json
@@ -7,15 +7,18 @@
   "categorizeByGroup": true,
   "groupOrder": [
     "Engine",
+    "Providers",
+    "Definers",
     "Controllers",
     "Buildable controllers",
     "Sub-controllers",
     "Actions",
     "*"
   ],
-  "name": "Insight",
-  "readme": "none",
-  "entryPoints": ["../../src/insight.index.ts"],
+  "name": "SSR Commerce",
+  "readme": "../../COMMERCEREADME.md",
+  "entryPoints": ["../../../headless-react/src/ssr-commerce/index.ts"],
   "gitRevision": "master",
-  "plugin": ["../../typedoc/dist/index.js"]
+  "plugin": ["../../dist/index.js"],
+  "projectDocuments": ["../../documents/headless-react/*.md"]
 }
