+++ packages/shared-typedoc/configs/headless-react/ssr.typedoc.json
@@ -7,15 +7,16 @@
   "categorizeByGroup": true,
   "groupOrder": [
     "Engine",
+    "Definers",
     "Controllers",
     "Buildable controllers",
     "Sub-controllers",
     "Actions",
     "*"
   ],
-  "name": "Commerce",
-  "readme": "none",
-  "entryPoints": ["../../src/commerce.index.ts"],
+  "name": "SSR Search",
+  "readme": "../../SEARCHREADME.md",
+  "entryPoints": ["../../../headless-react/src/ssr/index.ts"],
   "gitRevision": "master",
-  "plugin": ["../../typedoc/dist/index.js"]
+  "plugin": ["../../dist/index.js"]
 }
