+++ packages/shared-typedoc/configs/headless/index.typedoc.json
@@ -7,18 +7,15 @@
   "categorizeByGroup": true,
   "groupOrder": [
     "Engine",
-    "Providers",
-    "Definers",
     "Controllers",
     "Buildable controllers",
     "Sub-controllers",
     "Actions",
     "*"
   ],
-  "name": "SSR Commerce",
-  "readme": "../COMMERCEREADME.md",
-  "entryPoints": ["../../src/ssr-commerce/index.ts"],
+  "name": "Search",
+  "readme": "none",
+  "entryPoints": ["../../../headless/src/index.ts"],
   "gitRevision": "master",
-  "plugin": ["../../typedoc/dist/index.js"],
-  "projectDocuments": ["../documents/*.md"]
+  "plugin": ["../../dist/index.js"]
 }
