+++ packages/shared-typedoc/configs/headless/recommendation.typedoc.json
@@ -7,16 +7,15 @@
   "categorizeByGroup": true,
   "groupOrder": [
     "Engine",
-    "Definers",
     "Controllers",
     "Buildable controllers",
     "Sub-controllers",
     "Actions",
     "*"
   ],
-  "name": "SSR Commerce",
+  "name": "Recommendation",
   "readme": "none",
-  "entryPoints": ["../../src/ssr-commerce.index.ts"],
+  "entryPoints": ["../../../headless/src/recommendation.index.ts"],
   "gitRevision": "master",
-  "plugin": ["../../typedoc/dist/index.js"]
+  "plugin": ["../../dist/index.js"]
 }
