+++ packages/shared-typedoc/configs/headless/ssr-commerce.typedoc.json
@@ -14,9 +14,9 @@
     "Actions",
     "*"
   ],
-  "name": "SSR Search",
+  "name": "SSR Commerce",
   "readme": "none",
-  "entryPoints": ["../../src/ssr.index.ts"],
+  "entryPoints": ["../../../headless/src/ssr-commerce.index.ts"],
   "gitRevision": "master",
-  "plugin": ["../../typedoc/dist/index.js"]
+  "plugin": ["../../dist/index.js"]
 }
