+++ packages/shared-typedoc/lib/formatTypeDocToolbar.d.ts
@@ -1,0 +1,1 @@
+export declare function formatTypeDocToolbar(): void;
