+++ packages/shared-typedoc/lib/formatTypeDocToolbar.ts
