+++ packages/shared-typedoc/lib/insertAtomicSearchBox.ts
