+++ packages/shared-typedoc/lib/insertBetaNote.d.ts
@@ -1,0 +1,1 @@
+export declare function insertBetaNote(): void;
