+++ packages/shared-typedoc/lib/insertCoveoLogo.ts
