+++ packages/shared-typedoc/lib/insertCustomComments.d.ts
@@ -1,0 +1,2 @@
+import { Context, DeclarationReflection } from 'typedoc';
+export declare function insertCustomComments(this: undefined, _ctx: Context, refl: DeclarationReflection): void;
