+++ packages/shared-typedoc/lib/insertMetaTags.d.ts
@@ -1,0 +1,1 @@
+export declare function insertMetaTags(): void;
