+++ packages/shared-typedoc/lib/insertSiteHeaderBar.d.ts
@@ -1,0 +1,1 @@
+export declare function insertSiteHeaderBar(assetsPath: string): void;
