+++ packages/shared-typedoc/lib/insertSiteHeaderBar.ts
