+++ packages/shared-typedoc/package.json
@@ -1,0 +1,12 @@
+{
+  "name": "shared-typedoc",
+  "type": "module",
+  "files": [
+    "dist"
+  ],
+  "devDependencies": {
+    "@types/react": "18.3.20",
+    "@types/react-dom": "18.3.0",
+    "typedoc": "^0.28.3"
+  }
+}
